Fixes for aliasing for a small number of UV pixels
If the number of pixels is too small (< 5 for example) across the
aperture, the aperture structure itself is aliased and including a large
amount of padding will not alleviate the issue. The code now doubles the
number of pixels present if it's < 10px across, which seems to work for
my tests.

The `if` condition might need to be changed to a `while` condition, but
that change can be made later if necessary.