Start of weighted parang support in scripts/plumber.py
The --parang argument no longer accepts multiple values. If a single
value is passed, a beam at that parang is generated. To generated
weighted parang average, the --parang-file argument must be used, which
is a list of parangs and their associated weighs.

This can be used to also calculate the arithmetic mean by simply
assigning an equal weight to all the parallactic angles, making the
support for multiple --parang values redundant.