Support for weight parang averaging in place
This completes the skeleton features introduced in the previous commit.
The parang rotation is done via `np.rotate` since it's quicker than
`ia.rotate` and can be done in-memory.  The rotation and averaging are
done in the padded aperture domain, to keep the image sizes smaller than the
final output regridded images. Doing it on the zero-padded apertures
also ensures that edge effects due to the rotation are avoided.

A future commit will introduce unit tests to ensure that `ia.rotate` and
`np.rotate` do the same thing to an input image.