Changed directory structure to recommended
Moved the code into the `src` directory and the tests into the `tests`
directory. The `plumber-data` submodule was also moved into `tests`
since the intention is that the data in there will primarily be used for
tests.

Initial tests of `pyproject.toml` have worked with `pip install`.
Compatibility with other methods of installing (poetry etc.,) have not
been tested yet.