Initial implementation of image plane correction
Added image plane correction functionality - it's very rough around the
edges and needs to be refactored + tested, but this is the initial
implementation.

It requires each individual stokes plane to be passed in as a separate
image. In a future commit, this requirement will be relaxed, to allow
both inividual stokes plane images as well as a Stokes cube (single
frequency), and eventually will support Stokes spectral cubes.