Create parser module - for CSV files
The code to parse the CSV files and images already exists, but are
primarily contained within the zernikeBeam class. They do not belong
there, so I've started with moving the CSV file parsing into it's own
class, and will follow with the image parsing.

This also opens up the ability to auto-discover the location of the CSV
file, since that will require knowledge of which telescope made the
input image and selecting accordingly. Currently that logic gets setup
after initializing the zernikeBeam class, which requires an input CSV
- so it becomes a chicken and egg problem unless the image parser is
broken out into it's own class.