Bugfix to ensure correct PB sizes
Two bugs from the previous commit :
(a) --frequency was set to a flag rather than a float, so passing
a further argument would raise an error
(b) The `get_npix_aperture` function within `zernike.py` re-opened the
template image to read in the frequency. This would generate the wrong
number of pixels for the aperture, leading to a scaling issue. Now it
uses the frequency passed in from the `initialize` function.